requirements.txt
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
yfinance==0.2.28
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
praw==7.7.1
newspaper3k==0.2.8
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
litellm==1.20.3
openai==1.6.1
anthropic==0.8.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
cvxpy==1.4.1
# Optional: Uncomment if you want to use MongoDB/Redis for data storage
# pymongo==4.6.0
# redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
schedule==1.2.0
tqdm==4.66.1 